Add ignore file
Not sure if this will help anyone else, but on my end it makes it so my
`ag` CLI ignores the build directory.

The build directory seems somehow not to be entirely generated. If you
remove it, run `hugo` and run git status you will see there are things
remvoed. Odd.