Comment out unused code
Also added a version to the end of the links to prevent the following warning in the console:
Resource interpreted as Stylesheet but transferred with MIME type text/plain: "https://unpkg.com/driver.js/dist/driver.min.css".

In case the code will be used again, the warning should be gone.