Retire Visafinder
visafinder.tw has long hosted the gold card quick qualification
check. It has worked well, but always looked terrible, and the
server its on sometimes goes down for long periods.

Now the official folks have adopted the idea, and it's hosted on
a site with a very high uptime, and dedicated resources to keep
it up to date. Therefore, I think it's time to retire visafinder.

This patch changes the links from visafinder to
https://goldcard.nat.gov.tw/en/apply/step-1/

and updates the "Do I Qualify" link to point to our own qualification
section instead of the external site.

fixes #188