[WIP] Add Community Leadership Team Charter
This commit seperates the community page into its own directory
in the though that there will be additional content in the future,
such as a code of conduct/how to contribute, and other community-related
content.

It also adds a Community Leadership Team Charter, which is currently under
discussion. This patch should not be merged until consensus is reached.