Update health-insurance.md
The original link  to the NHI page about calculating premiums for 'special' cases is broken and returns an error page. I'm not sure where the original link was supposed to point, so as a temporary fix I changed the link to the English language version of the NHI site map.